chore(backend): move createAutoFormResponse from edusharing to common util
