+++ src/backend/edusharing/route-handlers.ts
@@ -3,7 +3,7 @@ import jwt from 'jsonwebtoken'
 import * as t from 'io-ts'
 import { readEnvVariable } from '../read-env-variable'
 import urlJoin from 'url-join'
-import { createAutoFormResponse } from './create-auto-form-response'
+import { createAutoFormResponse } from '../util/create-auto-form-response'
 import { verifyJwt } from './verify-jwt'
 import { createJWKSResponse } from './create-jwks-response'
 import { signJwtWithBase64Key } from './sign-jwt'
