+++ src/edusharing-mock/server.ts
@@ -7,7 +7,7 @@ import { v4 as uuid_v4 } from 'uuid'
 import * as jose from 'jose'
 import urlJoin from 'url-join'
 import { readEnvVariable } from '../backend/read-env-variable'
-import { createAutoFormResponse } from '../backend/edusharing/create-auto-form-response'
+import { createAutoFormResponse } from '../backend/util/create-auto-form-response'
 
 const editorUrl = readEnvVariable('EDITOR_URL')
